fix: some refactor for testing, added pytest to reqs
